// src/main/resources/webapp/src/app/common/model/system/menuinfo.ts
export class MenuInfo {
    menuname: string;

    icon: string;

    parentname: string;

    url: string;

    routerlink: string;

    isparent: string;

    description: string;

    sort: string;

    hideflag: string;

    reserved1: string;

    reserved2: string;

    reserved3: string;

    chinese: string;

    english: string;

    korean: string;

    parent: MenuInfo;
}